clippy: fix pedantic warnings [2] (#466)
* cargo: disable "pendantic::items_after_statements" warnings
* cargo: disable "pendantic::unused_self" warnings
* cargo: disable "pendantic::match_same_arms" warnings
* cargo: disable "pendantic::if_not_else" warnings
* cargo: fix "pendantic::manual_let_else" warnings
* cargo: fix "pendantic::needless_pass_by_value" warnings
* cargo: fix "pendantic::unnecessary_wraps" warnings
* cargo: disable "pendantic::struct_excessive_bools" warnings
* cargo: disable "pendantic::match_wildcard_for_single_variants" warnings
* cargo: disable "pendantic::struct_field_names" warnings
* cargo: fix "pendantic::doc_markdown" warnings
* cargo: fix "pendantic::trivially_copy_pass_by_ref" warnings
* cargo: fix "pendantic::from_iter_instead_of_collect" warnings
* cargo: fix "pendantic::linkedlist" warnings
* cargo: fix "pendantic::match_bool" warnings
* cargo: disable "pendantic::fn_params_excessive_bools" warnings
* cargo: fix "pendantic::should_panic_without_expect" warnings
* cargo: disable "pendantic::too_many_lines" warnings
* cargo: disable "pendantic::similar_names" warnings
* cargo: fix "pendantic::unused_async" warnings
* cargo: fix "pendantic::needless_pass_by_value" warnings [2]
* cargo: fix "pendantic::match_same_arms" warnings
* cargo: fix "pedantic::redundant_closure_for_method_calls" warnings
* cargo: fix "pedantic::cast_precision_loss" warnings
* cargo: fix "pedantic::cast_sign_loss" warnings
* cargo: fix "pedantic::cast_possible_truncation" warnings
* cargo: activate pedantic ruleset with exceptions
* DEVELOPER.md: add note about clippy configuration

+++ src/app/cli/src/commands/push_command.rs
@@ -289,8 +289,8 @@ impl SyncListener for PrettySyncProgress {
                         .progress_chars("#>-");
                     pb.set_style(style);
                     pb.set_prefix(format!("({} > {})", self.remote_ref, self.local_ref));
-                    pb.set_length(stats.src_estimated.metadata_blocks_read as u64);
-                    pb.set_position(stats.src.metadata_blocks_read as u64);
+                    pb.set_length(stats.src_estimated.metadata_blocks_read);
+                    pb.set_position(stats.src.metadata_blocks_read);
                     pb.enable_steady_tick(Duration::from_millis(100));
                     pb
                 }
@@ -306,8 +306,8 @@ impl SyncListener for PrettySyncProgress {
                         .progress_chars("#>-");
                     pb.set_style(style);
                     pb.set_prefix(format!("({} > {})", self.remote_ref, self.local_ref));
-                    pb.set_length(stats.dst_estimated.bytes_written as u64);
-                    pb.set_position(stats.dst.bytes_written as u64);
+                    pb.set_length(stats.dst_estimated.bytes_written);
+                    pb.set_position(stats.dst.bytes_written);
                     pb.enable_steady_tick(Duration::from_millis(100));
                     pb
                 }
@@ -322,8 +322,8 @@ impl SyncListener for PrettySyncProgress {
                         .progress_chars("#>-");
                     pb.set_style(style);
                     pb.set_prefix(format!("({} > {})", self.remote_ref, self.local_ref));
-                    pb.set_length(stats.dst_estimated.metadata_blocks_written as u64);
-                    pb.set_position(stats.dst.metadata_blocks_written as u64);
+                    pb.set_length(stats.dst_estimated.metadata_blocks_written);
+                    pb.set_position(stats.dst.metadata_blocks_written);
                     pb.enable_steady_tick(Duration::from_millis(100));
                     pb
                 }
@@ -336,24 +336,20 @@ impl SyncListener for PrettySyncProgress {
             SyncStage::ReadMetadata => {
                 state
                     .progress
-                    .set_length(stats.src_estimated.metadata_blocks_read as u64);
-                state
-                    .progress
-                    .set_position(stats.src.metadata_blocks_read as u64);
+                    .set_length(stats.src_estimated.metadata_blocks_read);
+                state.progress.set_position(stats.src.metadata_blocks_read);
             }
             SyncStage::TransferData => {
-                state
-                    .progress
-                    .set_length(stats.dst_estimated.bytes_written as u64);
-                state.progress.set_position(stats.dst.bytes_written as u64);
+                state.progress.set_length(stats.dst_estimated.bytes_written);
+                state.progress.set_position(stats.dst.bytes_written);
             }
             SyncStage::CommitBlocks => {
                 state
                     .progress
-                    .set_length(stats.dst_estimated.metadata_blocks_written as u64);
+                    .set_length(stats.dst_estimated.metadata_blocks_written);
                 state
                     .progress
-                    .set_position(stats.dst.metadata_blocks_written as u64);
+                    .set_position(stats.dst.metadata_blocks_written);
             }
         }
     }
